+++ src/main/java/com/github/jlangch/venice/impl/types/VncBigInteger.java
@@ -292,6 +292,11 @@ public class VncBigInteger extends VncNumber {
     }
 
     @Override
+    public float toJavaFloat() {
+        return value.floatValue();
+    }
+
+    @Override
     public BigInteger toJavaBigInteger() {
         return value;
     }