+++ src/main/java/com/github/jlangch/venice/impl/types/VncLong.java
@@ -289,6 +289,11 @@ public class VncLong extends VncNumber {
     }
 
     @Override
+    public float toJavaFloat() {
+        return value;
+    }
+
+    @Override
     public BigInteger toJavaBigInteger() {
         return BigInteger.valueOf(value);
     }